requirements.txt
Django>=5.2.4
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
djangorestframework-simplejwt>=5.5.0
gunicorn>=23.0.0
mysqlclient>=2.2.7
python-dotenv>=1.1.1